fix fixed duplicate releases
If a developer makes a release, then either forgets to push the tag
or decides they don't want to, and makes another release, the old
code would resume numbering yielding weird releases.

The new variant uses a regexp to find the latest release (this month)
and go from there. It has a check to ensure a bad tag doesn't make it
into the repo and we start releasing at a future time.